Modified the pizza.html link inside the homepage, as well as also added new information to the pizza.html page
On this commit, a few changes where made to the whole repository, as a
first step, a security measure was added to the relative link of the
pizza.html page in the index.html file. This was fixed by adding a the
target and rel tags to the href attribute. Secondly, the information
inside the pizza.html page was increased, including new paragraphs and
sorted and unsorted lists that represented the Steps and the ingredients
needed for its creation, as well as a <img> attribute that showed the
final result of the dish