Added two more recipe pages and sorted the links on the homepage on an unordered list
During this commit, two more recipes where added to the webpage, the
hotdog and coffee files followed the exact same structure as the pizza
recipe page, which consisted on the title of the document, a photo of
the final result of the dish, as well as a description and some
ingredients and steps that need to be follow in order to create the
dish. Also, the index.html file was modified, adding an unordered list
to give some structure to the relative links.